Lazy create socket for Statsd client, add BatchClient class
Statsd client used to create a socket on instanciation, but now uses
lazy techniques to defer creation of the socket to later, when it's
needed. (Also renamed socket attr to _socket to imply it's not part of
        the public API).
Added a new BatchClient class (not completed yet), to buffer the
requests and send them in batches.
A client object can be used as a context manager to create a BatchClient
object that shares the configuration and resource of the original
client, to avoid using unnecessary system resources.