Add TCPClient using a MixIn, rearrange class hierarchy in client
Using Abstract base classe for Clients, now BatchClient is not an
instance of Client anymore, because it did not make sense that the
BatchClient inherit the method "batch_client".
Also the batch client created from the client uses the same socket as
the client. To keep track of clients and their sockets (and close the
        sockets only if no one is using them), introduced a

_SharedSocket class that decorates the socket object, delegating socket
related operations to the original socket object but provides some
helper methods to keep track of socket usage and status.