Disallow chaning of remoet address (host, port).
For UDP client it was fine to change the address, however in TCP client
changing the address required reconnecting (or doing nothing if the
        address really did not change). This was making the code more
complicated, adding methods I did not like to add (_on_address_change
        event, etc.). Also made it more difficult to handle in multiple
threads, meaning the operation was not thread safe. To make it thread
safe I had to modify the AutoClosingSharedSocket thread safe which meant
I had to use locks and impor threading in there, which I did not wanted
to add all this complexity, import extra module in the package (where
        UDP client is defined and most likly is the target of most
        imports), just to be able to change the address after the client
object is created. I don't think changing the address after creating the
client is such an important feature to do all these just to support
that. So I'll drop the feature instead.