Fix broken tests on Jython and PyPy
The behavior of Python VMs for garbage collection is different amongs
supported implementations.
Explicitly ask for a garbage collection in tests when removing a client
and checking if the socket for the batch client is connected.
However Jython was still being a bit not cooperative, even calling the
garbage collection explicitly apparently did not call __del__, so
I skipped testing that shared socket will close when no more client is
using it for Jython.