Clean AutoClosingSharedSocket close method
Removed shutting down the socket since its mostly going to be used for
UDP which is no connected, and in general the protocol is unidirectional
send only, so there is no need to shutdown the socket and closing it
should be enough.

Also call the close() method of the object itself in __del__ instead of
closing the socket directly, to keep 1 single way to close the socket.