Simplify threaded TCP client
Removing closing the client explicitly. When the object is destroyed it
will automatically close and there is no need to provide an explicit
method for this.

Also removed joining the requests queue since that might block the main
thread (client of the API) which might not be desired.
If an application really needs to make sure all the queued metrics are
sent they could extend this class, join the queue and use that client
instead.