Improve threaded client by using a closed() event, add close() to it
Since a reference to the client object is avaialable on the thread that
sends requests, the client will not be garbage collected unless closed()
    explicitly.
I'm not sure how much this might cause troubles. I'm still thinking if
such feature to have a threaded client is that much helpful at all.
Will decide this later.