Add codecov to Travis, run tests with codecoverage
Also fix skipping TCP integration tests with PyPy.
I thought on CPython containers Travis won't have PyPY installed, but
that's not true. So need to find another way to skip running the TCP
integration tests with PyPy (this is only temporary until the issue
the test has with PyPy is fixed).