Use processes in TCP integration to start the server instead of threads
This fixes the issue that we had with PyPy keeping the test processing
running. So instead of starting a new dummy server in a separate thread,
we spawn a new process and run the server there, and instead of using
lists to collect stats requests for the server, we send the requests
back to the test process.