Add typing for statsdclient and batch client
Also overwrite the flush() method available in BatchClientMixIn in the
BatchClient because the MixIn class is using attributes from it's
extending classes which is not correct so it'll need to be removed.
But since TCPBatchClient still depends on this method from the
BatchClietMixIn, we'll remove it in a future commit.