Add client.timer.StopWatch class and allow setting rates for Timer
So the a Timer may have a default sample rate (instead of always using 1
as default), but each method that sends metrics may override the default
sample rate.

Assert valid values for sample rates.

Add StopWatch to convinently send timing metrics from a given starting
reference. This will help users to send metrics relative to a given poin
in time. This also will be used as the context manager for timing calls.