Clients create a stopwatch (configured using the client itself)
So that users can access to a configured stopwatch directly from the
clients (because we assume Clients are mostly to be instanciated rather
than other helper tools).

This also provides an nice feature for the clients to measure the
duration of a with block (since Stopwatch objects are context managers).

Tests for clients show how this new feature can be used to time a with
block.